test(remove): testing remove command

+++ src/Services/ComponentRemover.php
@@ -3,27 +3,42 @@
 
 namespace Sheaf\Cli\Services;
 
+use Illuminate\Console\Command;
 use Illuminate\Support\Facades\File;
-use Symfony\Component\Console\Output\ConsoleOutput;
-
+use Illuminate\Console\Concerns\InteractsWithIO;
+use Laravel\Prompts\Output\ConsoleOutput;
+use Symfony\Component\Console\Input\StringInput;
 
 class ComponentRemover
 {
     protected $output;
     protected $componentName;
 
-    public function __construct()
+    public function __construct(protected $command)
     {
-        $this->output = new ConsoleOutput();
     }
 
     public function remove($name)
     {
         $this->componentName = $name;
 
+        $isExists = $this->checkComponentExistence();
+
+        if(!$isExists) {
+            $this->message("Component is not installed in this project.");
+            return Command::FAILURE;
+        }
+
         $this->deleteComponentFiles();
 
         $this->cleaningSheafLock($name);
+    }
+
+    protected function checkComponentExistence()
+    {
+
+        return File::exists(resource_path("views/components/ui/{$this->componentName}")) ||
+            File::exists(resource_path("views/components/ui/{$this->componentName}.blade.php"));
     }
 
     protected function deleteComponentFiles()
@@ -32,7 +47,7 @@ class ComponentRemover
 
         if (File::isDirectory($componentDirectory)) {
             File::deleteDirectory($componentDirectory);
-            $this->message("+ Deleted directory: $componentDirectory");
+            $this->message("+ Deleted directory: resources/views/components/ui/{$this->componentName}");
         }
 
         $componentFile = resource_path("views/components/ui/{$this->componentName}.blade.php");
@@ -136,6 +151,6 @@ class ComponentRemover
 
     protected function message(string $message)
     {
-        $this->output->writeln("<fg=green>$message</fg=green>");
+        $this->command->info("<fg=green>$message</fg=green>");
     }
 }
